feat: 변하지 않는 데이터를 @Component, @PostConstruct 어노테이션을 사용해 빈으로 등록(#58)

+++ src/main/java/reviewme/be/util/repository/LabelRepository.java
@@ -5,7 +5,7 @@ import reviewme.be.util.entity.Label;
 
 import java.util.List;
 
-public interface LabelRepository extends JpaRepository<Label, Long> {
+public interface LabelRepository extends JpaRepository<Label, Integer> {
 
     List<Label> findByResumeIsNull();
 }
